requirements.txt
anyio==3.6.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
ch-orm==0.2.3
charset-normalizer==3.1.0
clickhouse-driver==0.2.5
clickhouse-sqlalchemy==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-auth-ldap==4.1.0
django-ninja==0.21.0
django-rest-framework==0.1.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
infi.clickhouse-orm==2.1.3
inflection==0.5.1
iso8601==1.1.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-ldap==3.4.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scipy==1.10.1
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.15
wheel==0.38.4
wincertstore==0.2
xgboost==1.7.4
xlwt==1.3.0
